+++ src/main/java/com/newspring/delivery/controllers/UsersController.java
@@ -1,9 +1,10 @@
 package com.newspring.delivery.controllers;
 
 import com.newspring.delivery.dto.common.OnlyStatusResponse;
-import com.newspring.delivery.dto.options_with_user.*;
-import com.newspring.delivery.dto.options_with_user.AdvanceOrderFiltersDto;
-import com.newspring.delivery.entities.*;
+import com.newspring.delivery.dto.optionsDto.usersDto.AddUserRequest;
+import com.newspring.delivery.dto.optionsDto.usersDto.GetAllRolesResponse;
+import com.newspring.delivery.dto.optionsDto.usersDto.UserWithRoleResponse;
+import com.newspring.delivery.entities.user.ChangeRoleOnUser;
 import com.newspring.delivery.mappers.UserMapper;
 import com.newspring.delivery.services.UsersService;
 import lombok.RequiredArgsConstructor;
@@ -63,22 +64,6 @@ public class UsersController {
         }
     }
 
-    @GetMapping("/portion")
-    public AdvanceOrdersResponse allAdvancedOrderResponse(AdvanceOrderFiltersDto advancedOrder) {
-        try {
-            log.info("response : {}",
-                    userMapper.toAdvancedOrders(usersService.advancedOrderSearch(userMapper.toAdvanceOrdersFilters(advancedOrder)))
-                    );
-            return userMapper.toAdvancedOrders(usersService.advancedOrderSearch(userMapper.toAdvanceOrdersFilters(advancedOrder)));
-        } catch (Exception e) {
-            AdvanceOrdersResponse response = new AdvanceOrdersResponse();
-            response.setStatus("ERROR");
-            response.setError(e.getMessage());
-            log.info("response : {}", e.getMessage(), e);
-            return response;
-        }
-    }
-
     @GetMapping("/part")
     public UserWithRoleResponse allUserWithRole(
             @RequestParam(required = false) Long role,
@@ -95,61 +80,11 @@ public class UsersController {
             return response;
         }
     }
-
-
-    @PostMapping("/create")
-    public OnlyStatusResponse createOrder(@RequestBody Order order) {
-        OnlyStatusResponse res = new OnlyStatusResponse();
-        try {
-            usersService.createOrder(order);
-            res.setStatus(OnlyStatusResponse.Status.OK);
-            res.setMessage(" -> заявка созданна");
-        } catch (Exception e) {
-            log.error("ERROR create order in Controller {}", e.getMessage(), e);
-            res.setStatus(OnlyStatusResponse.Status.FAIL);
-            res.setMessage(e.getMessage());
-        }
-        return res;
-    }
-
-    @PutMapping("/change")
-    public OnlyStatusResponse ChangeOrderRequest(@RequestBody ChangeOrder order) {
-        OnlyStatusResponse res = new OnlyStatusResponse();
-        try {
-            usersService.changeOrder(order);
-            res.setStatus(OnlyStatusResponse.Status.OK);
-            res.setMessage(" -> заявка обновлена");
-        } catch (Exception e) {
-            log.error("ERROR change order in Controller {}", e.getMessage(), e);
-            res.setStatus(OnlyStatusResponse.Status.FAIL);
-            res.setMessage(e.getMessage());
-        }
-        return res;
-    }
-
-    @DeleteMapping("/remove")
-    public OnlyStatusResponse removeOrder(@RequestBody DeleteOrderRequest deleteOrder) {
-        OnlyStatusResponse res = new OnlyStatusResponse();
-        try {
-            usersService.removeOrder(deleteOrder);
-            res.setStatus(OnlyStatusResponse.Status.OK);
-            res.setMessage(" -> заявка удалена");
-        } catch (Exception e) {
-            log.error("ERROR remove order in Controller {}", e.getMessage(), e);
-            res.setStatus(OnlyStatusResponse.Status.FAIL);
-            res.setMessage(e.getMessage());
-        }
-        return res;
-    }
 }
 // Запросы for Bash:
 // 1) curl -XPOST http://localhost:8080/users/add -H"Content-Type:application/json" -d'{"login":"Dima","password":"369852","roleId":2,"phone":"222111"}'
 // 2) curl -XGET http://localhost:8080/users/roles
 // Запросы в идеи
 // 4) curl -XPOST http://localhost:8080/users/up -H"Content-Type:application/json" -d"{\"id\":2,\"roleId\":2}" для консоли идеи.Разница во внешних кавычках.
-// 5) curl -XPOST http://localhost:8080/users/create -H"Content-Type:application/json" -d"{\"authorUserId\":1,\"executorUserId\":null,\"price\":150,\"name\":\"Marlboro Gold\",\"description\":\"medium nicotine cigarettes\",\"address\":\"Essentuki, Kalinina 2\",\"statusId\":1}"
-// 6) curl -XPOST http://localhost:8080/users/change -H"Content-Type:application/json" -d"{\"authorUserId\":2,\"name\":\"Milk\",\"description\":\"white milk\",\"address\":\"Kislovodsk, Vernadsky 5\",\"statusId\":1}"
-// 7) curl -XDELETE http://localhost:8080/users/remove -H "Content-Type:application/json" -d"{\"orderId\":5}"
 //  запросы из браузера
 // 3)  http://localhost:8080/users/part?role=2&login=A
-// 8)  http://localhost:8080/users/portion?name=ffe&description=ack&address=Kislo
