+++ src/main/java/com/newspring/delivery/dto/optionsDto/ordersDto/AdvanceOrderFiltersDto.java
@@ -1,4 +1,4 @@
-package com.newspring.delivery.dto.options_with_user;
+package com.newspring.delivery.dto.optionsDto.ordersDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
