+++ src/main/java/com/newspring/delivery/dto/optionsDto/ordersDto/AdvancedOrderDto.java
@@ -1,4 +1,4 @@
-package com.newspring.delivery.entities;
+package com.newspring.delivery.dto.optionsDto.ordersDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class AdvanceOrder {
+public class AdvancedOrderDto {
     private Long orderId;
     private Double price;
     private String name;
