+++ src/main/java/com/newspring/delivery/dto/optionsDto/simple_dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.newspring.delivery.dto.options_with_user.simple_dto;
+package com.newspring.delivery.dto.optionsDto.simple_dto;
 
 import lombok.Data;
