+++ src/main/java/com/newspring/delivery/dto/optionsDto/usersDto/UserWithRoleResponse.java
@@ -1,4 +1,4 @@
-package com.newspring.delivery.dto.options_with_user;
+package com.newspring.delivery.dto.optionsDto.usersDto;
 
 import lombok.Data;
