+++ src/main/java/com/newspring/delivery/entities/order/AdvanceOrder.java
@@ -1,4 +1,4 @@
-package com.newspring.delivery.dto.options_with_user;
+package com.newspring.delivery.entities.order;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class AdvancedOrderDto {
+public class AdvanceOrder {
     private Long orderId;
     private Double price;
     private String name;
